requirements.txt
attrs
python-morfeusz
requests
